:hammer: :sparkles:  Implement new GDP and population scenarios
Refactor the functions used to create the GDP and population scenarios and add new scenario options. This involved mainly portioning the code up into smaller functions, and reusing these functions for GDP and population scenario creation. The new scenarios that can now be created are based upon recent historic data from WDI (complemented with growth rates given by the James2019 dataset), short term projections until 2025 from IMF (for GDPpc) and WB (for pop) and reconverge to the original SSP GDPpc levels by 2100. After 2100 the scenarios follow a Bezier curve (with more or less arbirtray parameters) in such a way that by 2150 they're derivative is half of that in 2100. These new scenarios are used by default.
The SHAPE GDP scenarios are also implemented. These are scenarios that start with SDP_.
In order to further improve the clarity of the GDP and pop scenario construction, it is planned to move the associated functions into a separate package mrdirvers, in the near future.